+++ packages/demos/kitchen-sink/src/components/index.ts
@@ -2,17 +2,11 @@
 // Copyright 2022 DXOS.org
 //
 
-export * from './adapter';
-
-export * from './BoxContainer';
 export * from './CreateItemDialog';
 export * from './EchoBoard';
-export * from './EchoGraph';
 export * from './EchoGrid';
 export * from './EchoList';
 export * from './IconRadio';
-export * from './InvitationDialog';
 export * from './Layout';
 export * from './Searchbar';
-export * from './SelectionEditor';
 export * from './Theme';