+++ packages/sdk/react-client-testing/src/utils/strings.ts
@@ -1,0 +1,5 @@
+//
+// Copyright 2022 DXOS.org
+//
+
+export const capitalize = (text: string) => text.length ? text.charAt(0).toUpperCase() + text.slice(1) : text;