+++ packages/sdk/react-client-testing/stories/SelectionEditor.stories.tsx
@@ -8,11 +8,10 @@ import { Box } from '@mui/material';
 
 import { ClientProvider, ProfileInitializer } from '@dxos/react-client';
 
-import { execSelection, SelectionEditor } from '../src';
-import { defaultSelectionText, useTestParty } from './helpers';
+import { defaultSelectionText, execSelection, SelectionEditor, useTestParty } from '../src';
 
 export default {
-  title: 'KitchenSink/SelectionEditor'
+  title: 'react-client-testing/SelectionEditor'
 };
 
 const App = () => {