+++ packages/sdk/react-components/src/BoxContainer.tsx
@@ -29,8 +29,7 @@ interface StyledBoxProps extends BoxProps {
  * </Fullscreen>
  * ```
  */
-// TODO(burdon): Move to @dxos/react-components.
-export const BoxContainer = styled(Box, {
+export const BoxContainer: any = styled(Box, {
   shouldForwardProp: (prop) => prop !== 'expand' && prop !== 'column' && prop !== 'scrollX' && prop !== 'scrollY'
 })<StyledBoxProps>(({ expand, column, scrollX, scrollY }) => ({
   display: 'flex',